fix(adjustment): correct fade behaviour
