+++ module/item/item-attack.js
@@ -9,7 +9,7 @@ import {
     convertToDcFromItem,
     convertFromDC,
 } from "../utility/damage.js";
-import { defensivePowerAdjustmentMultiplier } from "../utility/adjustment.js";
+import { performAdjustment } from "../utility/adjustment.js";
 import { RequiresASkillRollCheck } from "../item/item.js";
 import { ItemAttackFormApplication } from "../item/item-attack-application.js";
 
@@ -1923,12 +1923,13 @@ async function _onApplyAdjustmentToSpecificToken(event, tokenId, damageData) {
     for (const reduce of reducesArray) {
         const targetActor = token.actor;
 
-        _performAdjustment(
+        performAdjustment(
             item,
             reduce,
             rawActivePointsDamage,
             actualActivePointDamage,
             item.system.XMLID === "TRANSFER",
+            false,
             targetActor,
         );
     }
@@ -1937,7 +1938,7 @@ async function _onApplyAdjustmentToSpecificToken(event, tokenId, damageData) {
         const targetActor =
             item.system.XMLID === "TRANSFER" ? item.actor : token.actor;
 
-        _performAdjustment(
+        performAdjustment(
             item,
             enhance,
             -rawActivePointsDamage,
@@ -1945,368 +1946,10 @@ async function _onApplyAdjustmentToSpecificToken(event, tokenId, damageData) {
                 ? -actualActivePointDamage
                 : -rawActivePointsDamage,
             item.system.XMLID === "TRANSFER",
+            false,
             targetActor,
         );
     }
-}
-
-function _determineCostPerActivePoint(
-    potentialCharacteristic,
-    powerTargetX,
-    targetActor,
-) {
-    // TODO: Not sure we need to use the characteristic here...
-    const powerInfo = getPowerInfo({
-        xmlid: potentialCharacteristic.toUpperCase(),
-        actor: targetActor,
-    });
-    return (
-        (powerTargetX
-            ? parseFloat(
-                  powerTargetX.system.activePoints / powerTargetX.system.value,
-              )
-            : parseFloat(powerInfo?.cost || powerInfo?.costPerLevel)) *
-        defensivePowerAdjustmentMultiplier(
-            potentialCharacteristic.toUpperCase(),
-            targetActor,
-        )
-    );
-}
-
-function _findExistingMatchingEffect(
-    item,
-    potentialCharacteristic,
-    powerTargetName,
-    targetActor,
-) {
-    // TODO: Variable Effect may result in multiple changes on same AE.
-    return targetActor.effects.find(
-        (effect) =>
-            effect.origin === item.uuid &&
-            effect.flags.target[0] ===
-                (powerTargetName?.uuid || potentialCharacteristic),
-    );
-}
-
-function _createCharacteristicAEChangeBlock(
-    potentialCharacteristic,
-    targetActor,
-) {
-    return {
-        // TODO: Why is this only characteristics for the key? What about powers?
-        // system.value is transferred to the actor, so not very useful,
-        // but we can enumerate via item.effects when determining value.
-        key: targetActor.system.characteristics?.[potentialCharacteristic]
-            ? `system.characteristics.${potentialCharacteristic}.max`
-            : "system.value",
-        value: 0,
-        mode: CONST.ACTIVE_EFFECT_MODES.ADD,
-    };
-}
-
-async function _createNewAdjustmentEffect(
-    item,
-    potentialCharacteristic,
-    powerTargetName,
-    rawActivePointsDamage,
-    isTransfer,
-    targetActor,
-) {
-    // Create new ActiveEffect
-    const activeEffect = {
-        name: `${item.system.XMLID} 0 ${
-            powerTargetName?.name || potentialCharacteristic
-        } (0 AP) [by ${item.actor.name}]`,
-        id: `${item.system.XMLID}.${item.id}.${
-            powerTargetName?.name || potentialCharacteristic
-        }`,
-        icon: item.img,
-        changes: [
-            _createCharacteristicAEChangeBlock(
-                potentialCharacteristic,
-                targetActor,
-            ),
-        ],
-        duration: {
-            seconds: 12,
-        },
-        flags: {
-            type: "adjustment",
-            activePoints: 0,
-            XMLID: item.system.XMLID,
-            source: targetActor.name,
-            target: [powerTargetName?.uuid || potentialCharacteristic],
-            key: potentialCharacteristic,
-        },
-        origin: item.uuid,
-    };
-
-    // If this is 5e then some characteristics are entirely calculated based on
-    // those. We only need to worry about 2 (DEX -> OCV & DCV and EGO -> OMCV & DMCV)
-    // as figured characteristics aren't adjusted.
-    if (targetActor.system.is5e) {
-        if (potentialCharacteristic === "dex") {
-            activeEffect.changes.push(
-                _createCharacteristicAEChangeBlock("ocv", targetActor),
-            );
-            activeEffect.flags.target.push("ocv");
-
-            activeEffect.changes.push(
-                _createCharacteristicAEChangeBlock("dcv", targetActor),
-            );
-            activeEffect.flags.target.push("dcv");
-        } else if (potentialCharacteristic === "ego") {
-            activeEffect.changes.push(
-                _createCharacteristicAEChangeBlock("omcv", targetActor),
-            );
-            activeEffect.flags.target.push("omcv");
-
-            activeEffect.changes.push(
-                _createCharacteristicAEChangeBlock("dmcv", targetActor),
-            );
-            activeEffect.flags.target.push("dmcv");
-        }
-    }
-
-    // DELAYEDRETURNRATE (loss for TRANSFER and all other adjustments) and DELAYEDRETURNRATE2 (gain for TRANSFER)
-    const dRR = item.findModsByXmlid("DELAYEDRETURNRATE");
-    const dRR2 = item.findModsByXmlid("DELAYEDRETURNRATE2");
-    const delayedReturnRate =
-        rawActivePointsDamage > 0 ? dRR : isTransfer ? dRR2 : dRR;
-    if (delayedReturnRate) {
-        switch (delayedReturnRate.OPTIONID) {
-            case "MINUTE":
-                activeEffect.duration.seconds = 60;
-                break;
-            case "FIVEMINUTES":
-                activeEffect.duration.seconds = 60 * 5;
-                break;
-            case "20MINUTES":
-                activeEffect.duration.seconds = 60 * 20;
-                break;
-            case "HOUR":
-                activeEffect.duration.seconds = 60 * 60;
-                break;
-            case "6HOURS":
-                activeEffect.duration.seconds = 60 * 60 * 6;
-                break;
-            case "DAY":
-                activeEffect.duration.seconds = 60 * 60 * 24;
-                break;
-            case "WEEK":
-                activeEffect.duration.seconds = 604800;
-                break;
-            case "MONTH":
-                activeEffect.duration.seconds = 2.628e6;
-                break;
-            case "SEASON":
-                activeEffect.duration.seconds = 2.628e6 * 3;
-                break;
-            case "YEAR":
-                activeEffect.duration.seconds = 3.154e7;
-                break;
-            case "FIVEYEARS":
-                activeEffect.duration.seconds = 3.154e7 * 5;
-                break;
-            case "TWENTYFIVEYEARS":
-                activeEffect.duration.seconds = 3.154e7 * 25;
-                break;
-            case "CENTURY":
-                activeEffect.duration.seconds = 3.154e7 * 100;
-                break;
-            default:
-                await ui.notifications.error(
-                    `DELAYEDRETURNRATE has unhandled option ${delayedReturnRate?.OPTIONID}`,
-                );
-        }
-    }
-
-    await targetActor.addActiveEffect(activeEffect);
-
-    return _findExistingMatchingEffect(
-        item,
-        potentialCharacteristic,
-        powerTargetName,
-        targetActor,
-    );
-}
-
-async function _performAdjustment(
-    item,
-    targetedPower,
-    rawActivePointsDamage,
-    activePointDamage,
-    isTransfer,
-    targetActor,
-) {
-    const targetName = targetedPower.toUpperCase();
-
-    // TODO: Not sure this should be o.name as would be the name of the power? Should it be using XMLID?
-    // TODO: Do we need to trim name? Presumably we shrink them when adding?
-    const powerTargetName = targetActor.items.find(
-        (o) => o.name.toUpperCase().trim() === targetName,
-    );
-    const potentialCharacteristic = powerTargetName
-        ? powerTargetName.system.XMLID
-        : targetName.toLowerCase();
-    const characteristicTarget =
-        targetActor.system.characteristics?.[potentialCharacteristic];
-
-    // A target we understand?
-    if (!characteristicTarget && !powerTargetName) {
-        // Can't find anything to link this against...meh. Might be caught by the validity check above.
-        return;
-    }
-
-    // Check for previous adjustment (i.e ActiveEffect) from same power against this target
-    // and calculate the total effect
-    const existingEffect = _findExistingMatchingEffect(
-        item,
-        potentialCharacteristic,
-        powerTargetName,
-        targetActor,
-    );
-    const activeEffect =
-        existingEffect ||
-        (await _createNewAdjustmentEffect(
-            item,
-            potentialCharacteristic,
-            powerTargetName,
-            rawActivePointsDamage,
-            isTransfer,
-            targetActor,
-        ));
-    let totalNewActivePoints =
-        activePointDamage + activeEffect.flags.activePoints;
-
-    // Clamp max change to the max allowed by the power.
-    // TODO: Combined effects may not exceed the largest maximum for a single target.
-    if (totalNewActivePoints < 0) {
-        totalNewActivePoints = Math.max(
-            totalNewActivePoints,
-            -item.system.maxAdjustment,
-        );
-    } else {
-        totalNewActivePoints = Math.min(
-            totalNewActivePoints,
-            item.system.maxAdjustment,
-        );
-    }
-
-    // Determine how many points of effect there are based on the cost
-    const costPerActivePoint = _determineCostPerActivePoint(
-        potentialCharacteristic,
-        powerTargetName,
-        targetActor,
-    );
-    const activePointsThatShouldBeAffected = Math.trunc(
-        totalNewActivePoints / costPerActivePoint,
-    );
-    const activePointAffectedDifference =
-        activePointsThatShouldBeAffected -
-        Math.trunc(activeEffect.flags.activePoints / costPerActivePoint);
-
-    activeEffect.changes[0].value =
-        activeEffect.changes[0].value - activePointAffectedDifference;
-
-    // If this is 5e then some characteristics are calculated (not figured) based on
-    // those. We only need to worry about 2: DEX -> OCV & DCV and EGO -> OMCV & DMCV.
-    // These 2 characteristics are always at indices 2 and 3
-    if (activeEffect.changes[1]) {
-        activeEffect.changes[1].value = Math.trunc(
-            activeEffect.changes[0].value / 3,
-        );
-    }
-    if (activeEffect.changes[2]) {
-        activeEffect.changes[2].value = Math.trunc(
-            activeEffect.changes[0].value / 3,
-        );
-    }
-
-    // Update the effect value(s)
-    activeEffect.name = `${item.system.XMLID} ${Math.abs(
-        activePointsThatShouldBeAffected,
-    )} ${powerTargetName?.name || potentialCharacteristic} (${Math.abs(
-        totalNewActivePoints,
-    )} AP) [by ${item.actor.name}]`;
-
-    activeEffect.flags.activePoints = totalNewActivePoints;
-
-    await activeEffect.update({
-        name: activeEffect.name,
-        changes: activeEffect.changes,
-        flags: activeEffect.flags,
-    });
-
-    // TODO: Pretty sure recovery isn't working as expected for defensive items
-
-    // TODO: Only needed for characteristics? How will this work for powers?
-    if (targetActor.system.characteristics?.[potentialCharacteristic]) {
-        const newValue =
-            targetActor.system.characteristics[potentialCharacteristic].value -
-            activePointAffectedDifference;
-        const changes = {
-            [`system.characteristics.${potentialCharacteristic}.value`]:
-                newValue,
-        };
-
-        if (targetActor.system.is5e && activeEffect.flags.target[1]) {
-            changes[
-                `system.characteristics.${activeEffect.flags.target[1]}.value`
-            ] = RoundFavorPlayerUp(newValue / 3);
-            changes[
-                `system.characteristics.${activeEffect.flags.target[2]}.value`
-            ] = RoundFavorPlayerUp(newValue / 3);
-        }
-
-        await targetActor.update(changes);
-    }
-
-    await _generateAdjustmentChatCard(
-        item,
-        activePointDamage,
-        activePointAffectedDifference,
-        totalNewActivePoints,
-        potentialCharacteristic,
-        targetActor,
-    );
-}
-
-async function _generateAdjustmentChatCard(
-    item,
-    activePointDamage,
-    activePointAffectedDifference,
-    totalActivePointEffect,
-    potentialCharacteristic, // TODO: Power?
-    targetActor,
-) {
-    const cardData = {
-        item: item,
-
-        adjustmentDamageRaw: activePointDamage,
-        adjustmentDamageThisApplication: activePointAffectedDifference,
-
-        adjustmentTarget: potentialCharacteristic.toUpperCase(),
-        adjustmentTotalActivePointEffect: totalActivePointEffect,
-
-        defense: 0, // TODO: FIXME:
-
-        targetActor: targetActor,
-    };
-
-    // render card
-    const template =
-        "systems/hero6efoundryvttv2/templates/chat/apply-adjustment-card.hbs";
-    const cardHtml = await renderTemplate(template, cardData);
-    const speaker = ChatMessage.getSpeaker({ actor: item.actor });
-
-    const chatData = {
-        user: game.user._id,
-        content: cardHtml,
-        speaker: speaker,
-    };
-
-    await ChatMessage.create(chatData);
 }
 
 async function _onApplySenseAffectingToSpecificToken(
